Splitting fixtures into different files
- Simplifies conftest
