+++ pytest.ini
@@ -3,4 +3,7 @@ filterwarnings =
     ignore:.* is deprecated:DeprecationWarning
     ignore:Call to deprecated create *.*
     ignore:.*distutils Version *.*
-addopts = --doctest-modules credoai
+addopts = 
+    --doctest-modules credoai
+    -p tests.fixtures.artifacts
+    -p tests.fixtures.datasets
