+++ tests/conftest.py
@@ -1,22 +1,13 @@
 import pickle
 
-from pytest import fixture
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.linear_model import LinearRegression, LogisticRegression
-from sklearn.model_selection import train_test_split
-
-from credoai.artifacts import ClassificationModel, RegressionModel, TabularData
-from credoai.datasets import fetch_creditdefault, fetch_testdata
-
 import pytest
+from connect.governance import Governance
 from pandas import DataFrame
+from pytest import fixture
 
-from credoai.artifacts import ComparisonData, TabularData
-from credoai.artifacts.model.comparison_model import DummyComparisonModel
-
+from credoai.artifacts import ClassificationModel, TabularData
 from credoai.lens import Lens
 
-from connect.governance import Governance
 
 ################################################
 ############ Lens init #########################
@@ -140,221 +131,10 @@ def init_lens_regression(
         model=regression_model,
         assessment_data=regression_assessment_data,
         training_data=regression_train_data,
-        governance=gov
+        governance=gov,
     )
 
     return my_pipeline, temp_file, gov
-
-
-################################################
-############ Artifacts fixtures ################
-################################################
-
-### Datasets definition ########################
-
-
-@fixture(scope="session")
-def binary_data():
-    train_data, test_data = fetch_testdata(False, 1, 1)
-    return {"train": train_data, "test": test_data}
-
-
-@fixture(scope="session")
-def continuous_data():
-    train_data, test_data = fetch_testdata(False, 1, 1, "continuous")
-    return {"train": train_data, "test": test_data}
-
-
-@fixture(scope="session")
-def credit_data():
-    data = fetch_creditdefault()
-    X = data["data"].iloc[0:100]
-    X = X.drop(columns=["SEX"])
-    y = data["target"].iloc[0:100].astype(int)
-    (
-        X_train,
-        X_test,
-        y_train,
-        y_test,
-    ) = train_test_split(X, y, random_state=42)
-    train_data = {"X": X_train, "y": y_train}
-    test_data = {"X": X_test, "y": y_test}
-    return {"train": train_data, "test": test_data}
-
-
-@fixture(scope="session")
-def ranking_fairness_data():
-    df = DataFrame(
-        {
-            "rankings": [1, 2, 3, 4, 5, 6, 7, 8],
-            "scores": [10, 8, 7, 6, 2, 2, 1, 1],
-            "sensitive_features": ["f", "f", "m", "m", "f", "m", "f", "f"],
-        }
-    )
-    return df
-
-
-@fixture(scope="session")
-def identity_verification_data():
-    source_subject_id = 4 * ["s0"] + 5 * ["s1"] + ["s2"]
-    source_subject_data_sample = 4 * ["s00"] + 3 * ["s10"] + 2 * ["s11"] + ["s20"]
-    target_subject_id = ["s1"] + ["s1", "s2", "s3"] * 2 + ["s2", "s3", "s3"]
-    target_subject_data_sample = (
-        ["s10"] + ["s11", "s20", "s30"] * 2 + ["s20"] + ["s30"] * 2
-    )
-    pairs = DataFrame(
-        {
-            "source-subject-id": source_subject_id,
-            "source-subject-data-sample": source_subject_data_sample,
-            "target-subject-id": target_subject_id,
-            "target-subject-data-sample": target_subject_data_sample,
-        }
-    )
-
-    subjects_sensitive_features = DataFrame(
-        {
-            "subject-id": ["s0", "s1", "s2", "s3"],
-            "gender": ["female", "male", "female", "female"],
-        }
-    )
-    return pairs, subjects_sensitive_features
-
-
-### Identity verification data/model artifacts ###
-
-
-@fixture(scope="session")
-def identity_verification_model():
-    similarity_scores = [31.5, 16.7, 20.8, 84.4, 12.0, 15.2, 45.8, 23.5, 28.5, 44.5]
-
-    credo_model = DummyComparisonModel(
-        name="face-compare", compare_output=similarity_scores
-    )
-    return credo_model
-
-
-@fixture(scope="session")
-def identity_verification_comparison_data(identity_verification_data):
-    pairs, subjects_sensitive_features = identity_verification_data
-    credo_data = ComparisonData(
-        name="face-data",
-        pairs=pairs,
-        subjects_sensitive_features=subjects_sensitive_features,
-    )
-    return credo_data
-
-
-### Ranking fairness data artifacts ############
-
-
-@fixture(scope="session")
-def ranking_fairness_assessment_data(ranking_fairness_data):
-    df = ranking_fairness_data
-    data = TabularData(
-        name="ranks",
-        y=df[["rankings", "scores"]],
-        sensitive_features=df[["sensitive_features"]],
-    )
-    return data
-
-
-### Classification model/data artifacts ########
-
-
-@fixture(scope="session")
-def classification_model(binary_data):
-    model = LogisticRegression(random_state=0)
-    model.fit(binary_data["train"]["X"], binary_data["train"]["y"])
-    credo_model = ClassificationModel("hire_classifier", model)
-    return credo_model
-
-
-@fixture(scope="session")
-def classification_assessment_data(binary_data):
-    test = binary_data["test"]
-    return TabularData(
-        name="assessment_data",
-        X=test["X"],
-        y=test["y"],
-        sensitive_features=test["sensitive_features"],
-    )
-
-
-@fixture(scope="session")
-def classification_train_data(binary_data):
-    train = binary_data["train"]
-    return TabularData(
-        name="training_data",
-        X=train["X"],
-        y=train["y"],
-        sensitive_features=train["sensitive_features"],
-    )
-
-
-### Regression model/data artifacts ############
-
-
-@fixture(scope="session")
-def regression_model(continuous_data):
-    model = LinearRegression()
-    model.fit(continuous_data["train"]["X"], continuous_data["train"]["y"])
-    credo_model = RegressionModel("skill_regressor", model)
-    return credo_model
-
-
-@fixture(scope="session")
-def regression_assessment_data(continuous_data):
-    test = continuous_data["test"]
-    return TabularData(
-        name="assessment_data",
-        X=test["X"],
-        y=test["y"],
-        sensitive_features=test["sensitive_features"],
-    )
-
-
-@fixture(scope="session")
-def regression_train_data(continuous_data):
-    train = continuous_data["train"]
-    return TabularData(
-        name="assessment_data",
-        X=train["X"],
-        y=train["y"],
-        sensitive_features=train["sensitive_features"],
-    )
-
-
-### Credit model/data artifacts ################
-
-
-@fixture(scope="session")
-def credit_classification_model(credit_data):
-    model = RandomForestClassifier()
-    model.fit(credit_data["train"]["X"], credit_data["train"]["y"])
-    credo_model = ClassificationModel("credit_default_classifier", model)
-    return credo_model
-
-
-@fixture(scope="session")
-def credit_assessment_data(credit_data):
-    test = credit_data["test"]
-    assessment_data = TabularData(
-        name="UCI-credit-default-test",
-        X=test["X"],
-        y=test["y"],
-    )
-    return assessment_data
-
-
-@fixture(scope="session")
-def credit_training_data(credit_data):
-    train = credit_data["train"]
-    training_data = TabularData(
-        name="UCI-credit-default-test",
-        X=train["X"],
-        y=train["y"],
-    )
-    return training_data
 
 
 ################################################
